fix(salary table): - change salary type from int to bigint

+++ migrations/1706102488133_create-table-salary.js
@@ -9,7 +9,7 @@ exports.up = (pgm) => {
       primaryKey: true,
     },
     salary: {
-      type: 'INT',
+      type: 'BIGINT',
       notNull: true,
     },
     owner: {
